move logTarget to context
