+++ commands_test.go
@@ -2,9 +2,7 @@ package main
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
-	"io"
 	"os"
 	"sort"
 	"strings"
@@ -15,83 +13,6 @@ import (
 	"github.com/creativeprojects/resticprofile/util/collect"
 	"github.com/stretchr/testify/assert"
 )
-
-func fakeCommands() *OwnCommands {
-	ownCommands := NewOwnCommands()
-	ownCommands.Register([]ownCommand{
-		{
-			name:              "first",
-			description:       "first first",
-			action:            firstCommand,
-			needConfiguration: false,
-		},
-		{
-			name:              "second",
-			description:       "second second",
-			action:            secondCommand,
-			needConfiguration: true,
-			flags: map[string]string{
-				"-f, --first":   "first flag",
-				"-s, --seccond": "second flag",
-			},
-		},
-		{
-			name:              "third",
-			description:       "third third",
-			action:            thirdCommand,
-			needConfiguration: false,
-			hide:              true,
-		},
-	})
-	return ownCommands
-}
-
-func firstCommand(_ io.Writer, _ commandContext) error {
-	return errors.New("first")
-}
-
-func secondCommand(_ io.Writer, _ commandContext) error {
-	return errors.New("second")
-}
-
-func thirdCommand(_ io.Writer, _ commandContext) error {
-	return errors.New("third")
-}
-
-func TestDisplayOwnCommands(t *testing.T) {
-	buffer := &strings.Builder{}
-	displayOwnCommands(buffer, commandContext{ownCommands: fakeCommands()})
-	assert.Equal(t, "  first   first first\n  second  second second\n", buffer.String())
-}
-
-func TestDisplayOwnCommand(t *testing.T) {
-	buffer := &strings.Builder{}
-	displayOwnCommandHelp(buffer, "second", commandContext{ownCommands: fakeCommands()})
-	assert.Equal(t, `Purpose: second second
-
-Usage:
-  resticprofile [resticprofile flags] [profile name.]second [command specific flags]
-
-Flags:
-  -f, --first    first flag
-  -s, --seccond  second flag
-
-`, buffer.String())
-}
-
-func TestIsOwnCommand(t *testing.T) {
-	assert.True(t, fakeCommands().Exists("first", false))
-	assert.True(t, fakeCommands().Exists("second", true))
-	assert.True(t, fakeCommands().Exists("third", false))
-	assert.False(t, fakeCommands().Exists("another one", true))
-}
-
-func TestRunOwnCommand(t *testing.T) {
-	assert.EqualError(t, fakeCommands().Run(&Context{request: Request{command: "first"}}), "first")
-	assert.EqualError(t, fakeCommands().Run(&Context{request: Request{command: "second"}}), "second")
-	assert.EqualError(t, fakeCommands().Run(&Context{request: Request{command: "third"}}), "third")
-	assert.EqualError(t, fakeCommands().Run(&Context{request: Request{command: "another one"}}), "command not found: another one")
-}
 
 func TestPanicCommand(t *testing.T) {
 	assert.Panics(t, func() {
