+++ context_test.go
@@ -7,6 +7,34 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestContextWithBinary(t *testing.T) {
+	ctx := &Context{
+		binary: "test",
+	}
+	ctx = ctx.WithBinary("test2")
+	assert.Equal(t, "test2", ctx.binary)
+}
+
+func TestContextWithCommand(t *testing.T) {
+	ctx := &Context{
+		command: "test",
+	}
+	ctx = ctx.WithCommand("test2")
+	assert.Equal(t, "test2", ctx.command)
+}
+
+func TestContextWithGroup(t *testing.T) {
+	ctx := &Context{
+		request: Request{
+			command: "test",
+			group:   "test",
+		},
+	}
+	ctx = ctx.WithGroup("test2")
+	assert.Equal(t, "test2", ctx.request.group)
+	assert.NotEmpty(t, ctx.request.command)
+}
+
 func TestContextWithProfile(t *testing.T) {
 	ctx := &Context{
 		request: Request{
